Smoothing up the build process (#63)
* Add prerequisites to book make target

In order to build the book, one needs to have some code and figures
generated. Adding this explicitly to the corresponding make target
allows to simply call `make book`, wait, and receive the final pdf file.

* Add prerequisites to slides make target

In order to build the slides, one needs to have some code and figures
generated. Adding this explicitly to the corresponding make target
allows to simply call `make genslides`, wait, and receive the final
pdf file.

* Fix LaTeX build error

The xcolor packages was provided with the non-existent option pdflatex.
This seems to be a rather common problem. The code is changed according to
the solution presented at StackOverflow:
https://tex.stackexchange.com/questions/713015/unknown-option-pdflatex-for-package-xcolor-edef

* Make the Makefile for the tutorial build as much as possible

---------

Co-authored-by: Oliver Keszöcze <olike@dtu.dk>